requirements.txt
fastapi==0.115.16
uvicorn==0.23.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.10
pydantic==2.10.5
python-dotenv~=1.0.1
